Move tools out of globals
