+++ src/all.scss
@@ -1,6 +1,7 @@
 // Settings
 
 @import "settings/all";
+@import "tools/all";
 
 @import "globals/common";
