+++ src/components/back-link/_back-link.scss
@@ -1,7 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/compatibility";
+@import "../../tools/all";
 
 @import "../../globals/helpers/media-queries";
 @import "../../globals/helpers/focusable";
