+++ src/components/breadcrumbs/_breadcrumbs.scss
@@ -1,9 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/compatibility";
-
-@import "../../globals/tools/ie8";
+@import "../../tools/all";
 
 @import "../../globals/helpers/clearfix";
 @import "../../globals/helpers/media-queries";
