+++ src/components/button/_button.scss
@@ -1,9 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/compatibility";
-@import "../../globals/tools/ie8";
-@import "../../globals/tools/iff";
+@import "../../tools/all";
 
 @import "../../globals/helpers/media-queries";
 @import "../../globals/helpers/focusable";
