+++ src/components/checkboxes/_checkboxes.scss
@@ -1,7 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/ie8";
+@import "../../tools/all";
 
 @import "../error-message/error-message";
 @import "../fieldset/fieldset";
