+++ src/components/date-input/_date-input.scss
@@ -1,8 +1,7 @@
 @import "../../settings/all";
+@import "../../tools/all";
 
 @import "../../globals/helpers/clearfix";
-@import "../../globals/tools/exports";
-@import "../../globals/tools/iff";
 
 @import "../error-message/error-message";
 @import "../input/input";
