+++ src/components/error-summary/_error-summary.scss
@@ -1,8 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/compatibility";
-@import "../../globals/tools/iff";
+@import "../../tools/all";
 
 @import "../../globals/helpers/media-queries";
 @import "../../globals/helpers/spacing";
