+++ src/components/fieldset/_fieldset.scss
@@ -1,7 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/iff";
+@import "../../tools/all";
 
 @import "../../globals/helpers/clearfix";
 @import "../../globals/helpers/spacing";
