+++ src/components/file-upload/_file-upload.scss
@@ -1,7 +1,7 @@
 @import "../../settings/all";
+@import "../../tools/all";
 
 @import "../../globals/helpers/focusable";
-@import "../../globals/tools/exports";
 
 @import "../error-message/error-message";
 @import "../hint/hint";
