+++ src/components/hint/_hint.scss
@@ -1,9 +1,8 @@
 @import "../../settings/all";
+@import "../../tools/all";
 
 @import "../../globals/helpers/typography";
 @import "../../globals/helpers/media-queries";
-
-@import "../../globals/tools/exports";
 
 @include govuk-exports("hint") {
   .govuk-hint {
