+++ src/components/radios/_radios.scss
@@ -1,8 +1,7 @@
 @import "../../settings/all";
+@import "../../tools/all";
 
 @import "../../globals/helpers/clearfix";
-@import "../../globals/tools/exports";
-@import "../../globals/tools/ie8";
 
 @import "../error-message/error-message";
 @import "../fieldset/fieldset";
