+++ src/components/table/_table.scss
@@ -1,8 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/iff";
-@import "../../globals/tools/px-to-em";
+@import "../../tools/all";
 
 @import "../../globals/helpers/media-queries";
 @import "../../globals/helpers/spacing";
