+++ src/components/textarea/_textarea.scss
@@ -1,8 +1,8 @@
 @import "../../settings/all";
+@import "../../tools/all";
 
 @import "../../globals/helpers/focusable";
 @import "../../globals/helpers/spacing";
-@import "../../globals/tools/exports";
 
 @import "../error-message/error-message";
 @import "../hint/hint";
