+++ src/components/warning-text/_warning-text.scss
@@ -1,7 +1,5 @@
 @import "../../settings/all";
-
-@import "../../globals/tools/exports";
-@import "../../globals/tools/iff";
+@import "../../tools/all";
 
 @import "../../globals/helpers/media-queries";
 @import "../../globals/helpers/spacing";
