+++ src/globals/_font-face.scss
@@ -1,4 +1,4 @@
-@import "tools/exports";
+@import "../tools/exports";
 
 @include govuk-exports("font-face") {
