+++ src/globals/core/_global-styles.scss
@@ -1,4 +1,4 @@
-@import "../tools/exports";
+@import "../../tools/exports";
 @import "links";
 @import "typography";
