+++ src/globals/helpers/_clearfix.scss
@@ -1,4 +1,4 @@
-@import "../tools/exports";
+@import "../../tools/exports";
 
 // Mixin to clear floats
 @mixin govuk-clearfix {
