+++ src/globals/helpers/_media-queries.scss
@@ -6,7 +6,7 @@
 // File is taken from https://github.com/sass-mq/sass-mq
 // Based on version 4.0.2
 
-@import "../tools/exports";
+@import "../../tools/exports";
 @import "../../settings/colours-palette";
 @import "../../settings/spacing";
