+++ src/tools/iff.test.js
@@ -2,7 +2,7 @@
 
 const util = require('util')
 
-const configPaths = require('../../../config/paths.json')
+const configPaths = require('../../config/paths.json')
 
 const sass = require('node-sass')
 const sassRender = util.promisify(sass.render)
@@ -15,7 +15,7 @@ const sassConfig = {
 describe('@function iff', () => {
   it('outputs if the condition is truthy', async () => {
     const sass = `
-      @import 'globals/tools/iff';
+      @import 'tools/iff';
 
       .foo {
         color: red iff(true, !important);
@@ -28,7 +28,7 @@ describe('@function iff', () => {
 
   it('does not output if the condition is falsey', async () => {
     const sass = `
-      @import 'globals/tools/iff';
+      @import 'tools/iff';
 
       .foo {
         color: red iff(false, !important);
